+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "resilient-rs"
-version = "0.2.1"
+version = "0.3.0"
 edition = "2024"
 
 description = "A Rust utility library for fault tolerance, including retry strategies, backoff mechanisms, and failure handling."
